Reduce code duplication and identify sign of more objects
