+++ pyccel/ast/utilities.py
@@ -16,18 +16,18 @@ from .core          import (AsName, Import, FunctionDef, FunctionCall,
                             Allocate, Duplicate, Assign, For, CodeBlock,
                             Concatenate, Module, PyccelFunctionDef)
 
-from .builtins      import (builtin_functions_dict,
+from .builtins      import (builtin_functions_dict, PythonLen, PythonAbs,
                             PythonRange, PythonList, PythonTuple)
 from .cmathext      import cmath_mod
 from .datatypes     import HomogeneousTupleType, PythonNativeInt, PrimitiveIntegerType
-from .internals     import PyccelInternalFunction, Slice
+from .internals     import PyccelInternalFunction, Slice, PyccelArrayShapeElement
 from .itertoolsext  import itertools_mod
 from .literals      import LiteralInteger, LiteralEllipsis, Nil
 from .operators     import PyccelUnarySub
 from .mathext       import math_mod
 from .sysext        import sys_mod
 
-from .numpyext      import (NumpyEmpty, NumpyArray, numpy_mod,
+from .numpyext      import (NumpyEmpty, NumpyArray, numpy_mod, NumpyAbs,
                             NumpyTranspose, NumpyLinspace)
 from .operators     import PyccelAdd, PyccelMul, PyccelIs, PyccelArithmeticOperator
 from .scipyext      import scipy_mod
@@ -779,3 +779,95 @@ def is_literal_integer(expr):
     return isinstance(a, (int, LiteralInteger)) or \
         (isinstance(a, PyccelUnarySub) and isinstance(a.args[0], (int, LiteralInteger))) or \
         (isinstance(a, Constant) and isinstance(a.dtype.primitive_type, PrimitiveIntegerType))
+
+#==============================================================================
+def get_expression_sign(expr):
+    """
+    Get an expression indicating the sign of the expression passed as an argument.
+
+    Get an expression indicating the sign of the expression passed as an argument.
+    If the expression represents a known literal integer then the value of the integer
+    is returned. If the expression is known to be positive then the return value
+    is 1. If the expression is known to be negative then the return value is -1.
+    If nothing is known about the sign of the expression the return value is None.
+    """
+    try:
+        sign = int(expr)
+    except TypeError:
+        if isinstance(expr, (PyccelArrayShapeElement, PythonLen, NumpyAbs, PythonAbs)):
+            sign = 1
+        elif isinstance(expr, UnarySub) and isinstance(expr.args[0], positive_types):
+            sign = -1
+        else:
+            sign = None
+
+    return sign
+
+#==============================================================================
+def get_new_slice_with_processed_arguments(_slice, array_size, allow_negative_index):
+    """
+    Create new slice with information collected from old slice and decorators.
+
+    Create a new slice where the original `start`, `stop`, and `step` have
+    been processed using basic simplifications, as well as additional rules
+    identified by the function decorators.
+
+    Parameters
+    ----------
+    _slice : Slice
+        Slice needed to collect (start, stop, step).
+
+    array_size : PyccelArrayShapeElement
+        Call to function size().
+
+    allow_negative_index : bool
+        True when the decorator allow_negative_index is present.
+
+    Returns
+    -------
+    Slice
+        The new slice with processed arguments (start, stop, step).
+    """
+    start = _slice.start
+    stop = _slice.stop
+    step = _slice.step
+
+    start_value = get_expression_sign(start)
+    stop_value = get_expression_sign(stop)
+    step_value = get_expression_sign(step)
+
+    # negative start and end in slice
+    if start_value and start_value < 0:
+        start = PyccelAdd(array_size, start, simplify = True)
+    elif start is not None and allow_negative_index and start_value is None:
+        start = IfTernaryOperator(PyccelLt(start, LiteralInteger(0)),
+                    PyccelAdd(array_size, start, simplify = True), start)
+
+    if stop_value and stop_value < 0:
+        stop = PyccelAdd(array_size, stop, simplify = True)
+    elif stop is not None and allow_negative_index and stop_value is None:
+        stop = IfTernaryOperator(PyccelLt(stop, LiteralInteger(0)),
+                    PyccelAdd(array_size, stop, simplify = True), stop)
+
+    # negative step in slice
+    if step_value and step_value < 0:
+        stop = PyccelAdd(stop, LiteralInteger(1), simplify = True) if stop is not None else LiteralInteger(0)
+        start = start if start is not None else PyccelMinus(array_size, LiteralInteger(1), simplify = True)
+
+    # variable step in slice
+    elif step and allow_negative_index and step_value is None:
+        if start is None:
+            start = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)),
+                LiteralInteger(0), PyccelMinus(array_size , LiteralInteger(1), simplify = True))
+
+        if stop is None:
+            stop = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)),
+                PyccelMinus(array_size, LiteralInteger(1), simplify = True), LiteralInteger(0))
+        else:
+            stop = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)),
+                stop, PyccelAdd(stop, LiteralInteger(1), simplify = True))
+
+    elif stop is not None:
+        stop = PyccelMinus(stop, LiteralInteger(1), simplify = True)
+
+    return Slice(start, stop, step)
