+++ pyccel/codegen/printing/ccode.py
@@ -46,7 +46,7 @@ from pyccel.ast.numpytypes import NumpyInt8Type, NumpyInt16Type, NumpyInt32Type,
 from pyccel.ast.numpytypes import NumpyFloat32Type, NumpyFloat64Type, NumpyComplex64Type, NumpyComplex128Type
 from pyccel.ast.numpytypes import NumpyNDArrayType, numpy_precision_map
 
-from pyccel.ast.utilities import expand_to_loops
+from pyccel.ast.utilities import expand_to_loops, get_expression_sign
 
 from pyccel.ast.variable import IndexedElement
 from pyccel.ast.variable import Variable
@@ -1388,7 +1388,7 @@ class CCodePrinter(CodePrinter):
                 #managing the Slice input
                 for i , ind in enumerate(inds):
                     if isinstance(ind, Slice):
-                        inds[i] = self._new_slice_with_processed_arguments(ind, PyccelArrayShapeElement(base, i),
+                        inds[i] = get_new_slice_with_processed_arguments(ind, PyccelArrayShapeElement(base, i),
                             allow_negative_indexes)
                     else:
                         inds[i] = Slice(ind, PyccelAdd(ind, LiteralInteger(1), simplify = True), LiteralInteger(1),
@@ -1441,66 +1441,6 @@ class CCodePrinter(CodePrinter):
             return f'(*{code})'
         else:
             return code
-
-    @staticmethod
-    def _new_slice_with_processed_arguments(_slice, array_size, allow_negative_index):
-        """
-        Create new slice with information collected from old slice and decorators.
-
-        Create a new slice where the original `start`, `stop`, and `step` have
-        been processed using basic simplifications, as well as additional rules
-        identified by the function decorators.
-
-        Parameters
-        ----------
-        _slice : Slice
-            Slice needed to collect (start, stop, step).
-
-        array_size : PyccelArrayShapeElement
-            Call to function size().
-
-        allow_negative_index : bool
-            True when the decorator allow_negative_index is present.
-
-        Returns
-        -------
-        Slice
-            The new slice with processed arguments (start, stop, step).
-        """
-        start = LiteralInteger(0) if _slice.start is None else _slice.start
-        stop = array_size if _slice.stop is None else _slice.stop
-
-        # negative start and end in slice
-        if isinstance(start, PyccelUnarySub) and isinstance(start.args[0], LiteralInteger):
-            start = PyccelMinus(array_size, start.args[0], simplify = True)
-        elif allow_negative_index and not isinstance(start, (LiteralInteger, PyccelArrayShapeElement)):
-            start = IfTernaryOperator(PyccelLt(start, LiteralInteger(0)),
-                            PyccelMinus(array_size, start, simplify = True), start)
-
-        if isinstance(stop, PyccelUnarySub) and isinstance(stop.args[0], LiteralInteger):
-            stop = PyccelMinus(array_size, stop.args[0], simplify = True)
-        elif allow_negative_index and not isinstance(stop, (LiteralInteger, PyccelArrayShapeElement)):
-            stop = IfTernaryOperator(PyccelLt(stop, LiteralInteger(0)),
-                            PyccelMinus(array_size, stop, simplify = True), stop)
-
-        # steps in slices
-        step = _slice.step
-
-        if step is None:
-            step = LiteralInteger(1)
-
-        # negative step in slice
-        elif isinstance(step, PyccelUnarySub) and isinstance(step.args[0], LiteralInteger):
-            start = PyccelMinus(array_size, LiteralInteger(1), simplify = True) if _slice.start is None else start
-            stop = LiteralInteger(0) if _slice.stop is None else stop
-
-        # variable step in slice
-        elif allow_negative_index and step and not isinstance(step, LiteralInteger):
-            og_start = start
-            start = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)), start, PyccelMinus(stop, LiteralInteger(1), simplify = True))
-            stop = IfTernaryOperator(PyccelGt(step, LiteralInteger(0)), stop, og_start)
-
-        return Slice(start, stop, step)
 
     def _print_PyccelArraySize(self, expr):
         arg = expr.arg
